RAC 443 Feat: 선배 상세 정보 조회 페이지 prefetchQuery적용, ErrorFallback컴포넌트 추가 (#303)
* feat: 선배 상세 페이지 에러 폴백 적용

* feat: 무한스크롤 시 api call 에러시 폴백 추가

* refactor: useErrorBoundary옵션 설정
