+++ package.json
@@ -35,6 +35,7 @@
     "react-calendar": "^4.7.0",
     "react-cookie": "6.1.1",
     "react-dom": "18.2.0",
+    "react-error-boundary": "^4.0.13",
     "react-hook-form": "^7.52.2",
     "styled-components": "6.1.0",
     "swiper": "11.0.5",
