+++ src/components/Card/ProfileCard/ProfileCard.tsx
@@ -51,9 +51,7 @@ function ProfileCard(props: ProfileCardProps) {
           <span>교수님</span>
         </div>
       </ProfileCardInfo>
-      <div id="profile-card-one-linear">
-        "{props.oneLinear ?? '한 줄 소개'}"
-      </div>
+      <div id="profile-card-one-linear">"{props.oneLinear ?? ''}"</div>
     </ProfileCardContainer>
   );
 }
